+++ module/documents/item.mjs
@@ -1648,7 +1648,6 @@ export default class Item5e extends Item {
    */
   async getChatData(htmlOptions = {}) {
     const data = this.toObject().system;
-    const labels = this.labels;
 
     // Rich text description
     data.description.value = await TextEditor.enrichHTML(data.description.value, {
@@ -1681,6 +1680,7 @@ export default class Item5e extends Item {
    */
   async rollAttack(options = {}) {
     const flags = this.actor.flags.sw5e ?? {};
+    const flagsCfg = CONFIG.SW5E.characterFlags;
     if (!this.hasAttack) throw new Error("You may not place an Attack Roll with this Item.");
     let title = `${this.name} - ${game.i18n.localize("SW5E.AttackRoll")}`;
 
@@ -1744,9 +1744,25 @@ export default class Item5e extends Item {
       }
     }
     // Flags
-    const elvenAccuracy =
-      (flags.elvenAccuracy && CONFIG.SW5E.characterFlags.elvenAccuracy.abilities.includes(this.abilityMod))
-      || undefined;
+    const elvenAccuracy = this._hasCharacterFlag("elvenAccuracy", this.abilityMod);
+
+    const advantageFlag = this._hasCharacterFlag([
+      "advantage.all",
+      "advantage.attack.all",
+      `advantage.attack.${this.abilityMod}`,
+      `advantage.attack.${this.system.actionType}`
+    ]);
+    let advantageHint = flagsCfg[advantageFlag]?.name ?? '';
+    if (flagsCfg[advantageFlag]?.condition) advantageHint += ` (${flagsCfg[advantageFlag].condition})`;
+
+    const disadvantageFlag = this._hasCharacterFlag([
+      "disadvantage.all",
+      "disadvantage.ability.save.all",
+      `disadvantage.attack.${this.abilityMod}`,
+      `disadvantage.attack.${this.system.actionType}`
+    ]);
+    let disadvantageHint = flagsCfg[disadvantageFlag]?.name ?? '';
+    if (flagsCfg[disadvantageFlag]?.condition) disadvantageHint += ` (${flagsCfg[disadvantageFlag].condition})`;
 
     // Compose roll options
     const rollConfig = foundry.utils.mergeObject(
@@ -1758,6 +1774,10 @@ export default class Item5e extends Item {
         flavor: title,
         elvenAccuracy,
         halflingLucky: flags.halflingLucky,
+        advantage: advantageFlag && !disadvantageFlag,
+        advantageHint,
+        disadvantage: disadvantageFlag && !advantageFlag,
+        disadvantageHint,
         dialogOptions: {
           width: 400,
           top: options.event ? options.event.clientY - 80 : null,
@@ -2216,9 +2236,16 @@ export default class Item5e extends Item {
         break;
       case "save":
         targets = this._getChatCardTargets(card);
+        const saveOptions = {
+          isForcePower: Object.keys(CONFIG.SW5E.powerSchoolsForce).includes(item.system.school),
+          isTechPower: Object.keys(CONFIG.SW5E.powerSchoolsTech).includes(item.system.school),
+          isPoison: item.system.consumableType === "poison",
+          dealsPoisonDmg: (item?.system?.damage?.parts ?? []).reduce(((acc, part) => acc || part[1] === "poison"), false),
+          dealsSonicDmg: (item?.system?.damage?.parts ?? []).reduce(((acc, part) => acc || part[1] === "sonic"), false),
+        };
         for (let token of targets) {
           const speaker = ChatMessage.getSpeaker({ scene: canvas.scene, token: token.document });
-          await token.actor.rollAbilitySave(button.dataset.ability, { event, speaker });
+          await token.actor.rollAbilitySave(button.dataset.ability, foundry.utils.mergeObject({ event, speaker }, saveOptions));
         }
         break;
       case "toolCheck":
@@ -2621,7 +2648,7 @@ export default class Item5e extends Item {
    * @private
    */
   _onCreateOwnedWeapon(data, isNPC) {
-    if ( !isNPC ) return;
+    if ( !isNPC ) return {};
     // NPCs automatically equip items.
     const updates = {};
     if ( !foundry.utils.hasProperty(data, "system.equipped") ) updates["system.equipped"] = true;